requirements.txt
streamlit
selenium
webdriver-manager
beautifulsoup4
requests
pydantic
pandas
pytest
python-dotenv
